リファクタリング(ページ分けた)
