+++ pubspec.yaml
@@ -26,7 +26,7 @@ dependencies:
   # Firebase
   firebase_core: ^0.4.5
   firebase_crashlytics: ^0.1.3+3
-  firebase_remote_config: ^0.3.1
+  # firebase_remote_config: ^0.3.1
   flutter:
     sdk: flutter
   share: ^0.6.4+3
@@ -37,6 +37,8 @@ dependencies:
   tutorial_coach_mark: ^0.3.0
   local_auth: ^0.6.2+3
   timeago: ^2.0.26
+  flutter_neumorphic: ^3.0.1
+  twitter_1user: ^1.0.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
